% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{calc_z}
\alias{calc_z}
\title{calc_z}
\usage{
calc_z(var, ...)
}
\arguments{
\item{var}{Variable to scale (create z score)}

\item{...}{additional arguments to pass to scale() (e.g. to change default
behaviour of scale and center arguments)}
}
\value{
Returns a vector with z scores calculated from called argument.
Typically assigned to a new column from the original data frame.
}
\description{
Calculates the z score of a given data frame column. Internal function
which is usually called as part of `create_z` function.
}
\examples{
## not run
## calc_z(lad_car)  Calculates z score for car ownership
}

