% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{create_z}
\alias{create_z}
\title{create_z}
\usage{
create_z(var)
}
\arguments{
\item{var}{A dataframe containing geography code, geography name, total
population, and population of interest (the Townsend variable) to perform
the z calculation on}
}
\value{
A tidy data frame containing geography code, geography name, and
z score
}
\description{
Creates a data frame containing the geography code, name, and z_score for
the variable requested
}
\details{
Expects at least the following columns (other columns are silenty dropped):
\enumerate{
 \item GEOGRAPHY_CODE
 \item GEOGRAPHY_NAME
 \item CELL_NAME
 \item OBS_VALUE
}

Data frames of this specific format can be obtained from Nomis Web.
}
\examples{
create_z(lad_car)
}
\seealso{
Nomis web: \url{http://nomisweb.co.uk}
}

