% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{get_shape}
\alias{get_shape}
\title{get_shape}
\usage{
get_shape(url, destfile, exdir, ...)
}
\arguments{
\item{url}{URL to obtain the shapefile from}

\item{destfile}{destination to download the zip archive to}

\item{exdir}{destination to unzip the shapefiles to}

\item{...}{extra arguments to pass to download.file (i.e. method = "wget")}
}
\value{
Returns an unzipped shapefile
}
\description{
Internal (not exported) function for obtaining and unziping shapefiles,
typically from edina.ac.uk
}
\details{
Typically used with `create_z()` function. Call for its side effects; i.e.
does not need to be assigned (`<-`)
}
\examples{
get_shape(shape_url, "inst/extdata/shape.zip", "inst/extdata")
}

