% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{prep_variable}
\alias{prep_variable}
\title{prep_variable}
\usage{
prep_variable(var)
}
\arguments{
\item{var}{The variable to prepare}
}
\value{

}
\description{
Prepares Townsend variables. Specifically removes last row if it contains
NAs, removes unneeded columns, and spreads the data to form a tidy data
frame
}
\details{
Expects the following columns:
1. GEOGRAPHY_CODE
1. GEOGRAPHY_NAME
1. CELL_NAME
1. OBS_VALUE

These are obtained from Nomis
}
\examples{
prep_variable(car)  # where car is a data frame object
}

