% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{calc_z}
\alias{calc_z}
\title{calc_z}
\usage{
calc_z(var, scale = TRUE, center = TRUE)
}
\arguments{
\item{var}{Variable to scale (create z score)}

\item{scale}{arguments passed to scale(). Defaults to TRUE}

\item{center}{arguments passed to scale(). Defaults to TRUE}
}
\value{

}
\description{
Calculates the z score of a given data frame variable. Usually called as
part of `create_z` function
}
\examples{
calc_z(lad_car) Calculates z score for car ownership
}

