% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{create_z}
\alias{create_z}
\title{create_z}
\usage{
create_z(var)
}
\arguments{
\item{var}{A dataframe containing geography code, geography name, total
population, and population of interest (the Townsend variable) to perform
the z calculation on}
}
\value{

}
\description{
Creates a data frame containing the geography code, name, and z_score for
the variable requested
}
\examples{
create_z(lad_car)
}

