% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{prep_variable}
\alias{prep_variable}
\title{prep_variable}
\usage{
prep_variable(var)
}
\arguments{
\item{var}{A data frame to tidy}
}
\value{
Returns a (tidy) data frame
}
\description{
Prepares Townsend variables from raw data obtained from Nomis web. Internal
function.
}
\details{
Removes last row if it contains NAs, removes unneeded columns, and spreads
the data to form a tidy data frame

Expects a data frame with at least the following columns (other columns can
be present and are silently dropped):
\enumerate{
 \item GEOGRAPHY_CODE
 \item GEOGRAPHY_NAME
 \item CELL_NAME
 \item OBS_VALUE
}
Data frames in this format can be obtained from Nomis
}
\examples{
## not run
## prep_variable(lad_car)  where lad_car is a data frame object
}
\seealso{
Nomis: \url{http://www.nomisweb.co.uk}
}

